Add permissions list to identity. Fixes #3430. Add read-only permission. Fixes #2109. Add CRUD permissions by entity type. Fixes #3435

+++ controller/internal/routes/authenticator_router.go
@@ -17,16 +17,18 @@
 package routes
 
 import (
+	"time"
+
 	"github.com/go-openapi/runtime/middleware"
+	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/edge-api/rest_management_api_server/operations/authenticator"
 	"github.com/openziti/foundation/v2/errorz"
 	"github.com/openziti/ziti/controller/env"
 	"github.com/openziti/ziti/controller/fields"
-	"github.com/openziti/ziti/controller/internal/permissions"
 	"github.com/openziti/ziti/controller/model"
+	"github.com/openziti/ziti/controller/permissions"
 	"github.com/openziti/ziti/controller/response"
 	"github.com/pkg/errors"
-	"time"
 )
 
 func init() {
@@ -46,36 +48,44 @@ func NewAuthenticatorRouter() *AuthenticatorRouter {
 
 func (r *AuthenticatorRouter) Register(ae *env.AppEnv) {
 	ae.ManagementApi.AuthenticatorDeleteAuthenticatorHandler = authenticator.DeleteAuthenticatorHandlerFunc(func(params authenticator.DeleteAuthenticatorParams, _ interface{}) middleware.Responder {
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Delete)
 		return ae.IsAllowed(r.Delete, params.HTTPRequest, params.ID, "", permissions.IsAdmin())
 	})
 
 	ae.ManagementApi.AuthenticatorDetailAuthenticatorHandler = authenticator.DetailAuthenticatorHandlerFunc(func(params authenticator.DetailAuthenticatorParams, _ interface{}) middleware.Responder {
-		return ae.IsAllowed(r.Detail, params.HTTPRequest, params.ID, "", permissions.IsAdmin())
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Read)
+		return ae.IsAllowed(r.Detail, params.HTTPRequest, params.ID, "", permissions.HasAdminReadOnly())
 	})
 
 	ae.ManagementApi.AuthenticatorListAuthenticatorsHandler = authenticator.ListAuthenticatorsHandlerFunc(func(params authenticator.ListAuthenticatorsParams, _ interface{}) middleware.Responder {
-		return ae.IsAllowed(r.List, params.HTTPRequest, "", "", permissions.IsAdmin())
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Read)
+		return ae.IsAllowed(r.List, params.HTTPRequest, "", "", permissions.HasAdminReadOnly())
 	})
 
 	ae.ManagementApi.AuthenticatorUpdateAuthenticatorHandler = authenticator.UpdateAuthenticatorHandlerFunc(func(params authenticator.UpdateAuthenticatorParams, _ interface{}) middleware.Responder {
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Update)
 		return ae.IsAllowed(func(ae *env.AppEnv, rc *response.RequestContext) { r.Update(ae, rc, params) }, params.HTTPRequest, params.ID, "", permissions.IsAdmin())
 	})
 
 	ae.ManagementApi.AuthenticatorCreateAuthenticatorHandler = authenticator.CreateAuthenticatorHandlerFunc(func(params authenticator.CreateAuthenticatorParams, _ interface{}) middleware.Responder {
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Create)
 		return ae.IsAllowed(func(ae *env.AppEnv, rc *response.RequestContext) { r.Create(ae, rc, params) }, params.HTTPRequest, "", "", permissions.IsAdmin())
 	})
 
 	ae.ManagementApi.AuthenticatorPatchAuthenticatorHandler = authenticator.PatchAuthenticatorHandlerFunc(func(params authenticator.PatchAuthenticatorParams, _ interface{}) middleware.Responder {
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Update)
 		return ae.IsAllowed(func(ae *env.AppEnv, rc *response.RequestContext) { r.Patch(ae, rc, params) }, params.HTTPRequest, params.ID, "", permissions.IsAdmin())
 	})
 
 	ae.ManagementApi.AuthenticatorReEnrollAuthenticatorHandler = authenticator.ReEnrollAuthenticatorHandlerFunc(func(params authenticator.ReEnrollAuthenticatorParams, _ interface{}) middleware.Responder {
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Update)
 		return ae.IsAllowed(func(ae *env.AppEnv, rc *response.RequestContext) {
 			r.ReEnroll(ae, rc, params)
 		}, params.HTTPRequest, params.ID, "", permissions.IsAdmin())
 	})
 
 	ae.ManagementApi.AuthenticatorRequestExtendAuthenticatorHandler = authenticator.RequestExtendAuthenticatorHandlerFunc(func(params authenticator.RequestExtendAuthenticatorParams, _ interface{}) middleware.Responder {
+		ae.InitPermissionsContext(params.HTTPRequest, permissions.Management, "authenticator", permissions.Update)
 		return ae.IsAllowed(func(ae *env.AppEnv, rc *response.RequestContext) {
 			r.RequestExtend(ae, rc, params)
 		}, params.HTTPRequest, params.ID, "", permissions.IsAdmin())
@@ -91,40 +101,60 @@ func (r *AuthenticatorRouter) Detail(ae *env.AppEnv, rc *response.RequestContext
 }
 
 func (r *AuthenticatorRouter) Create(ae *env.AppEnv, rc *response.RequestContext, params authenticator.CreateAuthenticatorParams) {
+	if !validateNonAdminAccess(ae, rc) {
+		return
+	}
+
 	Create(rc, rc, AuthenticatorLinkFactory, func() (string, error) {
-		authenticator, err := MapCreateToAuthenticatorModel(params.Authenticator)
+		authEntity, err := MapCreateToAuthenticatorModel(params.Authenticator)
 
 		if err != nil {
 			return "", err
 		}
 
-		return MapCreate(ae.Managers.Authenticator.Create, authenticator, rc)
+		return MapCreate(ae.Managers.Authenticator.Create, authEntity, rc)
 	})
 }
 
 func (r *AuthenticatorRouter) Delete(ae *env.AppEnv, rc *response.RequestContext) {
+	if !validateNonAdminAccess(ae, rc) {
+		return
+	}
+
 	DeleteWithHandler(rc, ae.Managers.Authenticator)
 }
 
 func (r *AuthenticatorRouter) Update(ae *env.AppEnv, rc *response.RequestContext, params authenticator.UpdateAuthenticatorParams) {
+	if !validateNonAdminAccess(ae, rc) {
+		return
+	}
+
 	Update(rc, func(id string) error {
 		return ae.Managers.Authenticator.Update(MapUpdateAuthenticatorToModel(params.ID, params.Authenticator), false, nil, rc.NewChangeContext())
 	})
 }
 
 func (r *AuthenticatorRouter) Patch(ae *env.AppEnv, rc *response.RequestContext, params authenticator.PatchAuthenticatorParams) {
+	if !validateNonAdminAccess(ae, rc) {
+		return
+	}
+
 	Patch(rc, func(id string, fields fields.UpdatedFields) error {
-		model := MapPatchAuthenticatorToModel(params.ID, params.Authenticator)
+		modelEntity := MapPatchAuthenticatorToModel(params.ID, params.Authenticator)
 
 		if fields.IsUpdated("password") {
 			fields.AddField("salt")
 		}
 
-		return ae.Managers.Authenticator.Update(model, false, fields.FilterMaps("tags"), rc.NewChangeContext())
+		return ae.Managers.Authenticator.Update(modelEntity, false, fields.FilterMaps("tags"), rc.NewChangeContext())
 	})
 }
 
 func (r *AuthenticatorRouter) ReEnroll(ae *env.AppEnv, rc *response.RequestContext, params authenticator.ReEnrollAuthenticatorParams) {
+	if !validateNonAdminAccess(ae, rc) {
+		return
+	}
+
 	id, err := rc.GetEntityId()
 
 	if err != nil {
@@ -150,6 +180,10 @@ func (r *AuthenticatorRouter) ReEnroll(ae *env.AppEnv, rc *response.RequestConte
 }
 
 func (r *AuthenticatorRouter) RequestExtend(ae *env.AppEnv, rc *response.RequestContext, params authenticator.RequestExtendAuthenticatorParams) {
+	if !validateNonAdminAccess(ae, rc) {
+		return
+	}
+
 	id, err := rc.GetEntityId()
 
 	if err != nil {
@@ -170,4 +204,30 @@ func (r *AuthenticatorRouter) RequestExtend(ae *env.AppEnv, rc *response.Request
 	}
 
 	rc.RespondWithEmptyOk()
+}
+
+func validateNonAdminAccess(ae *env.AppEnv, rc *response.RequestContext) bool {
+	if err := checkIfNonAdminIsModifyAdmin(ae, rc); err != nil {
+		rc.RespondWithError(err)
+		return false
+	}
+	return true
+}
+
+func checkIfNonAdminIsModifyAdmin(ae *env.AppEnv, rc *response.RequestContext) error {
+	if !rc.HasPermission(permissions.AdminPermission) {
+		id, err := rc.GetEntityId()
+		if err != nil {
+			log := pfxlog.Logger()
+			log.Error(err)
+			return err
+		}
+
+		if authEntity, _ := ae.Managers.Authenticator.Read(id); authEntity != nil {
+			if identityEntity, _ := ae.Managers.Identity.Read(authEntity.IdentityId); identityEntity != nil && identityEntity.IsAdmin {
+				return nonAdminNotAllowedError(errors.New("non-admin may not delete admin identity authenticators"))
+			}
+		}
+	}
+	return nil
 }
